move TestResourceIDs to resource/
