+++ core/src/test/scala/org/apache/spark/SparkConfSuite.scala
@@ -26,7 +26,7 @@ import scala.util.{Random, Try}
 import com.esotericsoftware.kryo.Kryo
 
 import org.apache.spark.resource.ResourceUtils._
-import org.apache.spark.TestResourceIDs._
+import org.apache.spark.resource.TestResourceIDs._
 import org.apache.spark.internal.config._
 import org.apache.spark.internal.config.History._
 import org.apache.spark.internal.config.Kryo._
