+++ core/src/test/scala/org/apache/spark/executor/CoarseGrainedExecutorBackendSuite.scala
@@ -34,7 +34,7 @@ import org.scalatest.mockito.MockitoSugar
 import org.apache.spark._
 
 import org.apache.spark.resource.ResourceUtils._
-import org.apache.spark.TestResourceIDs._
+import org.apache.spark.resource.TestResourceIDs._
 import org.apache.spark.TestUtils._
 import org.apache.spark.resource.{ResourceAllocation, ResourceInformation}
 import org.apache.spark.rpc.RpcEnv
