+++ core/src/test/scala/org/apache/spark/scheduler/TaskSetManagerSuite.scala
@@ -28,7 +28,7 @@ import org.mockito.invocation.InvocationOnMock
 
 import org.apache.spark._
 import org.apache.spark.resource.ResourceUtils._
-import org.apache.spark.TestResourceIDs._
+import org.apache.spark.resource.TestResourceIDs._
 import org.apache.spark.internal.Logging
 import org.apache.spark.internal.config
 import org.apache.spark.serializer.SerializerInstance
